fix OFT and OFTV2 deployment code and corresponding README.md instructions
The instructions for deploying OFTV2 in README.md did not work, as the deployment scripts refer
to ExampleOFT and ExampleOFTV2 contracts, which don't exist.  I changed these to OFTMock and
OFTV2Mock respectively.  Additionally, instructions failed to mention that the deployer *must*
set a minDstGasLimit in this version of code.  I added this instruction in and was able to send
tokens successfully.

Lastly, even though there is not an example for OFT, we failed to include a task to mintTokens(...).
This PR includes a small task to mint tokens so the OFT example can be tested.

Signed-off-by: Ryan Goulding <goulding@layerzerolabs.org>